# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fruits API",
        "description": "API documentation for the Fruits system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API server"
        }
    ],
    "paths": {
        "/auth": {},
        "/cards": {},
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng ký người dùng mới",
                "operationId": "ee65fb3f9c9eb5626a80236e13c1683d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nguyenvana@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng nhập người dùng",
                "operationId": "ab30188b420cffef6b7117d6e4decb92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nguyenvana@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng nhập thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Email hoặc mật khẩu không đúng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email hoặc mật khẩu không đúng"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng xuất người dùng",
                "operationId": "1b9d815a90eaf5d9476864adffeb387a",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng xuất thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/cards": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Lấy danh sách card",
                "operationId": "6b183c199a0477b18ceae55b1788a12f",
                "responses": {
                    "200": {
                        "description": "Danh sách card",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Card"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "Tạo mới card",
                "operationId": "b6049411123456119d22285a9be9859a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CardStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Card được tạo thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/cards/{id}": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Lấy thông tin chi tiết của một card",
                "operationId": "8b492d0de04c5abf87f6b7d66fc17f52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của card",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin chi tiết của card",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card không tìm thấy"
                    }
                }
            },
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "Cập nhật thông tin card",
                "operationId": "ad623be2cb056721844202f40f5fef75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của card",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CardUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Card được cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card không tìm thấy"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Card"
                ],
                "summary": "Xóa một card",
                "operationId": "1be8a6e1c9aa99c9ae5dc019e2e376a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của card",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card được xóa thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Card deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card không tìm thấy"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Lấy thông tin người dùng hiện tại",
                "operationId": "82b49a06ce04dff319bf4f8bc3ab561a",
                "responses": {
                    "200": {
                        "description": "Thông tin người dùng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/cards/random": {
            "get": {
                "tags": [
                    "Card - User"
                ],
                "summary": "Lấy ngẫu nhiên một card dựa trên tỷ lệ rarity",
                "operationId": "a7cd3341ba36d05414bb06134525423e",
                "responses": {
                    "200": {
                        "description": "Card ngẫu nhiên",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không có card nào phù hợp"
                    }
                }
            }
        },
        "/api/user/cards": {
            "get": {
                "tags": [
                    "Card - User"
                ],
                "summary": "Lấy danh sách card của người dùng",
                "operationId": "8f4bb1a4f5d087f0ae7011c3af773e79",
                "responses": {
                    "200": {
                        "description": "Danh sách card của người dùng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Card"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CardStoreRequest": {
                "required": [
                    "name",
                    "stars",
                    "rarity"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Dragon Fruit"
                    },
                    "stars": {
                        "type": "integer",
                        "example": 5
                    },
                    "description": {
                        "type": "string",
                        "example": "A rare and exotic fruit card."
                    },
                    "rarity": {
                        "type": "string",
                        "example": "Legendary"
                    },
                    "image_url": {
                        "description": "Upload image file",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object",
                "format": "multipart"
            },
            "CardUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Dragon Fruit"
                    },
                    "stars": {
                        "type": "integer",
                        "example": 5
                    },
                    "description": {
                        "type": "string",
                        "example": "A rare and exotic fruit card."
                    },
                    "rarity": {
                        "type": "string",
                        "example": "Legendary"
                    },
                    "image_url": {
                        "description": "Upload image file",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object",
                "format": "multipart"
            },
            "Card": {
                "title": "Card",
                "required": [
                    "id",
                    "name",
                    "stars",
                    "description",
                    "rarity",
                    "image_url"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dragon Fruit"
                    },
                    "stars": {
                        "type": "integer",
                        "example": 5
                    },
                    "description": {
                        "type": "string",
                        "example": "A rare and exotic fruit card."
                    },
                    "rarity": {
                        "type": "string",
                        "example": "Legendary"
                    },
                    "image_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/images/dragon_fruit.png"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "description": "ID của người dùng",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Tên của người dùng",
                        "type": "string",
                        "example": "Nguyen Van A"
                    },
                    "email": {
                        "description": "Email của người dùng",
                        "type": "string",
                        "format": "email",
                        "example": "nguyenvana@example.com"
                    },
                    "image_url": {
                        "description": "URL ảnh đại diện của người dùng",
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/images/nguyenvana.jpg"
                    },
                    "role": {
                        "description": "Vai trò của người dùng",
                        "type": "string",
                        "example": "user"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập token JWT vào ô Authorization",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Quản lý xác thực người dùng"
        },
        {
            "name": "Card",
            "description": "Quản lý card"
        },
        {
            "name": "Card - User",
            "description": "Quản lý card người dùng"
        }
    ]
}